+++ src/store.tsx
@@ -2,7 +2,7 @@ import { useDispatch } from 'react-redux';
 import { createStore, applyMiddleware } from 'redux';
 import thunkMiddleware from 'redux-thunk';
 import { composeWithDevTools } from 'redux-devtools-extension';
-import reducer from './redux/reducers/reducer';
+import reducer from './redux/reducers';
 
 const store = createStore(reducer, composeWithDevTools(applyMiddleware(thunkMiddleware)));
